pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speakx-notification-engine"
version = "1.0.0"
description = "AI-powered journey notification engine for SpeakX"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",
    "httpx>=0.26.0",
    "anthropic>=0.40.0",
    "python-statemachine>=2.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "aiosqlite>=0.19.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.setuptools.packages.find]
include = ["app*", "workers*"]

[tool.mypy]
python_version = "3.11"
strict = true
